[MS Connector] Refactor: add file.ts
Description
---
Split the activities.ts that had become too fat

Risk
---
na
